# src/resources/swagger-def.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Studies Service",
    "version": "1.0.0"
  },
  "security": [
    {
      "bearerAuth": ["STUDY-${programId}.WRITE"]
    }
  ],
  "paths": {
    "/health": {
      "get": {
        "tags": ["Health"],
        "summary": "get status of app",
        "responses": {
          "200": {
            "description": "app is healthy",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "500": {
            "description": "app is not healthy"
          }
        }
      }
    },
    "/studies": {
      "get": {
        "tags": ["Studies"],
        "summary": "Get current state of submission service for sample registration and clinical submission",
        "responses": {
          "200": {
            "description": "Boolean value of whether submission system is disabled or enabled"
          }
        }
      },
      "post": {
        "tags": ["Studies"],
        "summary": "Run dicitonary migration to update clinical dictionary (use dry-run-update to verify migration behavior first)",
        "description": "update and load a new schema version from lectern to be used.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreaetStudyReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "a dictioary migration object describing the migration run",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Studies"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/studies/submitters": {
      "post": {
        "tags": ["Studies"],
        "summary": "Get active registration",
        "description": "returns the current in progress registration for a program if any",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddSubmittersReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/GenericSuccess"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      },
      "delete": {
        "tags": ["Studies"],
        "summary": "Upload registration file",
        "description": "This will validate and save the new registration file overwriting any existing registration",
        "parameters": [
          {
            "in": "query",
            "name": "studyId",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "submitter",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/GenericSuccess"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    }
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "tags": [
    {
      "name": "Health",
      "description": "Service status monitoring"
    },
    {
      "name": "Studies",
      "description": "Studies management endpoints"
    }
  ],
  "components": {
    "responses": {
      "GenericSuccess": {
        "description": "Endpoint finished successfully"
      },
      "ServiceUnavailableError": {
        "description": "Endpoint service unavailable"
      },
      "UnauthorizedError": {
        "description": "Access token is missing or invalid"
      },
      "ServerError": {
        "description": "Server error"
      },
      "ForbiddenError": {
        "description": "Access token has incorrect scopes"
      },
      "NotFoundError": {
        "description": "Requested resource was not found"
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "CreaetStudyReq": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "studyId": {
            "type": "string"
          },
          "organization": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
      "AddSubmittersReq": {
        "type": "object",
        "properties": {
          "studyId": {
            "type": "string"
          },
          "submitters": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "Studies": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/Study"
        }
      },
      "Study": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "studyId": {
            "type": "string"
          },
          "organization": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "submitters": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "ServiceError": {
        "type": "object",
        "required": ["reason"],
        "properties": {
          "reason": {
            "type": "string"
          },
          "errorStudyId": {
            "type": "string"
          },
          "errorSubmitters": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      }
    }
  }
}